build: use Deno 1.4.6
This commit sets the version of Deno to be 1.4.6 instead of what ever
the latest version is. This could be made configurable.

The motivation for this changes is that the bundle created currently
does not work with 1.5.1 and the following error is generated:

error: Uncaught ReferenceError: __default is not defined
console.log(`Welcome to Deno(${__default(date)}:${__default(date)}) 🦕`)